+++ js/src/index.js
@@ -9,5 +9,5 @@
 // __webpack_public_path__ = document.querySelector('body').getAttribute('data-base-url') + 'nbextensions/jupyter-webrtc/';
 
 // Export widget models and views, and the npm package version number.
-module.exports = require("./webrtc.js");
-module.exports["version"] = require("../package.json").version;
+export * from "./jupyter-webrtc";
+// module.exports['version'] = require('../package.json').version;
