+++ js/src/jupyter-webrtc.js
@@ -1,0 +1,8 @@
+export * from "./streams/Audio";
+export * from "./streams/Camera";
+export * from "./streams/Image";
+export * from "./streams/Media";
+export * from "./streams/Recorder";
+export * from "./streams/Video";
+export * from "./streams/Webrtc";
+export * from "./streams/Widget";
